release 0.1.170 source code

+++ services/hss/v5/model/model_list_host_rasp_protect_history_info_response.go
@@ -9,10 +9,10 @@ import (
 // ListHostRaspProtectHistoryInfoResponse Response Object
 type ListHostRaspProtectHistoryInfoResponse struct {
 
-	// total number of dynamic WTPs
-	TotalNum *int64 `json:"total_num,omitempty"`
+	// **参数解释**: 总数 **取值范围**: 最小值0，最大值2147483647
+	TotalNum *int32 `json:"total_num,omitempty"`
 
-	// data list
+	// **参数解释**: 动态网页防篡改防护事件列表 **取值范围**: 最小值0，最大值200
 	DataList       *[]HostRaspProtectHistoryResponseInfo `json:"data_list,omitempty"`
 	HttpStatusCode int                                   `json:"-"`
 }
